+++ R/cached_data.R
@@ -1,3 +1,5 @@
+#' Cached ADSL
+#'
 #' Cached ADSL data generated with \code{seed = 1}
 #'
 #' @docType data
@@ -5,7 +7,8 @@
 #' @usage data(cadsl)
 "cadsl"
 
-
+#' Cached ADAE
+#'
 #' Cached ADAE data generated with \code{seed = 1}
 #'
 #' @docType data
@@ -13,6 +16,8 @@
 #' @usage data(cadae)
 "cadae"
 
+#' Cached ADAETTE
+#'
 #' Cached ADAETTE data generated with \code{seed = 1}
 #'
 #' @docType data
@@ -20,6 +25,8 @@
 #' @usage data(cadaette)
 "cadaette"
 
+#' Cached ADCM
+#'
 #' Cached ADCM data generated with \code{seed = 1}
 #'
 #' @docType data
@@ -27,55 +34,8 @@
 #' @usage data(cadcm)
 "cadcm"
 
-#' Cached ADLB data generated with \code{seed = 1}
+#' Cached ADEG
 #'
-#' @docType data
-#'
-#' @usage data(cadlb)
-"cadlb"
-
-#' Cached ADMH data generated with \code{seed = 1}
-#'
-#' @docType data
-#'
-#' @usage data(cadmh)
-"cadmh"
-
-#' Cached ADQS data generated with \code{seed = 1}
-#'
-#' @docType data
-#'
-#' @usage data(cadqs)
-"cadqs"
-
-#' Cached ADRS data generated with \code{seed = 1}
-#'
-#' @docType data
-#'
-#' @usage data(cadrs)
-"cadrs"
-
-#' Cached ADTTE data generated with \code{seed = 1}
-#'
-#' @docType data
-#'
-#' @usage data(cadtte)
-"cadtte"
-
-#' Cached ADVS data generated with \code{seed = 1}
-#'
-#' @docType data
-#'
-#' @usage data(cadvs)
-"cadvs"
-
-#' Cached ADEX data generated with \code{seed = 1}
-#'
-#' @docType data
-#'
-#' @usage data(cadex)
-"cadex"
-
 #' Cached ADEG data generated with \code{seed = 1}
 #'
 #' @docType data
@@ -83,9 +43,72 @@
 #' @usage data(cadeg)
 "cadeg"
 
+#' Cached ADEX
+#'
+#' Cached ADEX data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadex)
+"cadex"
+
+#' Cached ADLB
+#'
+#' Cached ADLB data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadlb)
+"cadlb"
+
+#' Cached ADMH
+#'
+#' Cached ADMH data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadmh)
+"cadmh"
+
+#' Cached ADQS
+#'
+#' Cached ADQS data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadqs)
+"cadqs"
+
+#' Cached ADRS
+#'
+#' Cached ADRS data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadrs)
+"cadrs"
+
 #' Cached ADTR data generated with \code{seed = 1}
 #'
 #' @docType data
 #'
 #' @usage data(cadtr)
 "cadtr"
+
+#' Cached ADTTE
+#'
+#' Cached ADTTE data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadtte)
+"cadtte"
+
+#' Cached ADVS
+#'
+#' Cached ADVS data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadvs)
+"cadvs"