+++ NEWS.md
@@ -1,8 +1,7 @@
 # random.cdisc.data 0.3.5
 
 * Added ADTR dataset from osprey.
-* Added 10 variables to ADAE dataset: AEOUT, AEACN, AESDTH, TRTEMFL, ANL01FL, SMQ01NAM,
-SMQ02NAM, SMQ01SC, SMQ02SC, CQ01NAM.
+* Added 10 variables to ADAE dataset: AEOUT, AEACN, AESDTH, TRTEMFL, ANL01FL, SMQ01NAM, SMQ02NAM, SMQ01SC, SMQ02SC, CQ01NAM.
 * Added WORS02FL (Worst Post-Baseline Observation) to ADEG dataset.
 * Added ADEG dataset.
 * Added 11 variables to ADLB dataset: BNRIND, SHIFT1, ATOXGR, BTOXGR, WOS01FL, ANL01FL, WGRHIFL, WGRLOFL, WGRHIVFL, WGRLOVFL, ONTRTFL.